cluster: add class to lazily enqueue similarity scoring tasks
One of the problems that seems to be facing the system is that
enqueueing a of tasks in advance takes up a lot of memory. So, I wrote
code to only enqueu as many tasks as there are workers. That way, new
tasks are only pushed when they can be worked on. Otherwise, the system
blocks. Hopefully, this makes the system have a smaller memory
footprint.