clustering: add option to save compatible patterns and input them
Add a feature to speed up grid search by adding a noop scorer and a flag to save
compatible patterns. The idea is to save all the compatible regexes and then pass them
the next time with the no-op scorer. This way, only compatible patterns are used, and
no string generation has to happen.